split query builder into separate file
