+++ lib/composer.js
@@ -1,11 +1,8 @@
 'use strict'
 const { isDeepStrictEqual } = require('node:util')
-const {
-  createEmptyObject,
-  unwrapSchemaType,
-  valueToArgumentString
-} = require('./gql-utils')
+const { createEmptyObject, unwrapSchemaType } = require('./gql-utils')
 const { fetchSubgraphSchema, makeGqlRequest } = require('./network')
+const { QueryBuilder } = require('./query-builder')
 const { isObject } = require('./utils')
 const {
   validateArray,
@@ -318,7 +315,7 @@ class Composer {
               type,
               field,
               fieldName: originalFieldName,
-              upstream: subgraph
+              subgraph
             })
 
             this.resolvers[typeName] ??= Object.create(null)
@@ -333,289 +330,92 @@ class Composer {
     }
   }
 
-  #createResolver ({ type, field, fieldName, upstream }) {
+  #createResolver ({ type, field, fieldName, subgraph }) {
     return async (parent, args, contextValue, info) => {
       const ctx = {
         path: [],
         followups: new Map(),
         result: { [fieldName]: null }
       }
-      // TODO(cjihrig): This should be done in a loop over info.fieldNodes.
       const node = info.fieldNodes[0]
       const schemaNode = type.fields.find((f) => {
         return f.name === node.name.value
       })
       const bareType = unwrapSchemaType(schemaNode.type)
       const objType = this.#types.get(bareType.name)
-      const { fieldMap } = objType
-      const query = {
+      const query = new QueryBuilder({
         node,
         schemaNode,
         path: [fieldName],
         type: objType,
         fields: [],
-        upstream,
+        subgraph,
         resolverName: info.fieldName,
         root: true,
         info,
-        adapter: null
-      }
-
-      for (const field of fieldMap.values()) {
-        query.fields.push(field)
-      }
-
-      ctx.followups.set(node, query)
-      await this.#buildQuery(query, ctx)
-
+        adapter: null,
+        types: this.#types
+      })
+      await this.#runQuery(ctx, query)
       return ctx.result[fieldName]
     }
   }
 
-  async #buildQuery (queryInfo, parentCtx) {
-    const ctx = {
-      path: parentCtx.path,
-      followups: new Map(),
-      result: parentCtx.result
-    }
-
-    const { info, node, resolverName, root, schemaNode, upstream } = queryInfo
-    const selectionSet = this.#buildSelectionSet(
-      queryInfo, ctx, node, schemaNode
-    )
-    const computedArgs = this.#buildArguments(queryInfo, ctx, node)
-    const query = `${info.operation.operation} { ${resolverName}${computedArgs} ${selectionSet} }`
-    console.log(' --> ' + query + '\n')
-    const data = await makeGqlRequest(query, upstream.server)
-
-    // TODO(cjihrig): Need to merge arrays properly.
-    let result = data[resolverName]
-    let mergeAt = ctx.result
-    let mergeParent = null
-
-    for (let i = 0; i < queryInfo.path.length; ++i) {
-      mergeParent = mergeAt
-      mergeParent[queryInfo.path[i]] ??= null
-      mergeAt = mergeAt[queryInfo.path[i]]
-    }
-
-    if (!root) {
-      if (Array.isArray(mergeAt) && !Array.isArray(result)) {
-        console.log('turning it into an array')
-        result = [result]
-      } else if (!Array.isArray(mergeAt) &&
-                 Array.isArray(result) &&
-                 result.length === 1) {
-        console.log('getting the first element of the array')
-        result = result[0]
-      }
-    }
-
-    // TODO(cjihrig): Need to perform merge with concept of owner.
-    if (mergeAt === null) {
-      mergeParent[queryInfo.path.at(-1)] = result
-      mergeAt = result
-    } else if (Array.isArray(result)) {
-      console.log('HERE!!!!!!!!')
-      console.log(mergeParent)
-      console.log(mergeAt)
-    } else if (isObject(result)) {
-      for (const [k, v] of Object.entries(result)) {
-        mergeAt[k] = v
-      }
-    } else {
-      mergeParent[queryInfo.path.at(-1)] = result
-    }
+  async #runQuery (ctx, query) {
+    const text = query.buildQuery(ctx)
+    const data = await makeGqlRequest(text, query.subgraph.server)
+    const merged = mergeResults(query, ctx.result, data)
 
     // TODO(cjihrig): Need to query the primary owner first.
     for (const followup of ctx.followups.values()) {
       const hasUnresolvedFields = followup.fields.some((f) => {
-        return !(f.schemaNode.name in mergeAt)
+        return !(f.schemaNode.name in merged)
       })
 
       if (!hasUnresolvedFields) {
         continue
       }
 
-      await this.#buildQuery(followup, ctx)
+      await this.#runQuery(ctx, new QueryBuilder(followup))
+    }
+  }
+}
+
+function mergeResults (query, partialResult, response) {
+  let result = response[query.resolverName]
+  let mergeAt = partialResult
+  let mergeParent = null
+
+  for (let i = 0; i < query.path.length; ++i) {
+    mergeParent = mergeAt
+    mergeParent[query.path[i]] ??= null
+    mergeAt = mergeAt[query.path[i]]
+  }
+
+  if (!query.root) {
+    if (Array.isArray(mergeAt) && !Array.isArray(result)) {
+      result = [result]
+    } else if (!Array.isArray(mergeAt) &&
+                Array.isArray(result) &&
+                result.length === 1) {
+      result = result[0]
     }
   }
 
-  #buildArguments (queryInfo, ctx, node) {
-    const length = node.arguments?.length ?? 0
-
-    if (length === 0) {
-      return ''
+  if (mergeAt === null) {
+    mergeParent[query.path.at(-1)] = result
+    mergeAt = result
+  } else if (Array.isArray(result)) {
+    // TODO(cjihrig): Need to merge arrays properly.
+  } else if (isObject(result)) {
+    for (const [k, v] of Object.entries(result)) {
+      mergeAt[k] = v
     }
-
-    const { info, adapter } = queryInfo
-
-    if (typeof adapter === 'function') {
-      let result = ctx.result
-
-      for (let i = 0; i < queryInfo.path.length; ++i) {
-        result = result[queryInfo.path[i]]
-      }
-
-      const args = adapter(result)
-      // TODO(cjihrig): Validate returned args object.
-      const mappedArgs = Object.keys(args).map((argName) => {
-        return `${argName}: ${args[argName]}`
-      })
-
-      return `(${mappedArgs.join(', ')})`
-    }
-
-    const mappedArgs = node.arguments.map((a) => {
-      const name = a.name.value
-      let value
-
-      if (a.value.kind === 'Variable') {
-        const varName = a.value.name.value
-        const varValue = info.variableValues[varName]
-        // const varDef = info.operation.variableDefinitions.find((v) => {
-        //   return v.variable.name.value === varName;
-        // });
-        // TODO(cjihrig): Use varDef to find strings.
-        const isString = false
-
-        if (typeof varValue === 'object') {
-          // TODO(cjihrig): Make this recursive and move to its own function.
-          const kvs = Object.keys(varValue).map((k) => {
-            let v = varValue[k]
-
-            if (typeof v === 'string') {
-              v = `"${v}"`
-            }
-
-            return `${k}: ${v}`
-          }).join(', ')
-
-          value = `{ ${kvs} }`
-        } else {
-          value = isString ? `"${varValue}"` : varValue
-        }
-      } else {
-        value = valueToArgumentString(a.value)
-      }
-
-      return `${name}: ${value}`
-    })
-
-    return `(${mappedArgs.join(', ')})`
+  } else {
+    mergeParent[query.path.at(-1)] = result
   }
 
-  #buildSelectionSet (queryInfo, ctx, node, schemaNode) {
-    const selections = node.selectionSet?.selections
-    const length = selections?.length ?? 0
-
-    if (length === 0) {
-      return ''
-    }
-
-    ctx.path.push(node.name.value)
-
-    const { info, upstream } = queryInfo
-    const bareType = unwrapSchemaType(schemaNode.type)
-    const type = this.#types.get(bareType.name)
-    const { fieldMap } = type
-    const set = []
-    let keyFields
-
-    for (let i = 0; i < length; ++i) {
-      const s = selections[i]
-
-      if (s.kind === 'FragmentSpread' || s.kind === 'InlineFragment') {
-        // TODO(cjihrig): The fragment probably only needs to be expanded if it
-        // is defined in the router. If it is defined by the subgraph server, it
-        // should know how to handle it.
-        const fragment = s.kind === 'FragmentSpread'
-          ? info.fragments[s.name.value]
-          : s
-
-        for (let i = 0; i < fragment.selectionSet.selections.length; ++i) {
-          const fragNode = fragment.selectionSet.selections[i]
-          let value = fragNode.name.value
-
-          if (fragNode.selectionSet) {
-            // TODO(cjihrig): Need to make this whole branch equivalent to the else branch.
-            value += ` ${this.#buildSelectionSet(queryInfo, ctx, fragNode, schemaNode)}`
-          }
-
-          set.push(value)
-        }
-      } else {
-        let value = s.name.value
-        const field = fieldMap.get(value)
-        const selectionSchemaNode = field?.schemaNode
-
-        // Some nodes won't have a schema representation. For example, __typename.
-        if (selectionSchemaNode) {
-          if (!field.subgraphs.has(upstream)) {
-            let existing = ctx.followups.get(node)
-
-            if (existing === undefined) {
-              // TODO(cjihrig): Double check the next line.
-              const nextUpstream = Array.from(field.subgraphs)[0]
-              const typeName = type.schemaNode.name
-              // TODO(cjihrig): Throw if this doesn't exist.
-              const entity = nextUpstream.entities[typeName]
-              const {
-                adapter,
-                foreignKeyFields,
-                primaryKeyFields,
-                referenceListResolverName: resolverName
-              } = entity
-
-              keyFields = primaryKeyFields ?? foreignKeyFields
-
-              existing = {
-                node,
-                schemaNode,
-                path: ctx.path.slice(),
-                type,
-                fields: [],
-                resolverName,
-                root: false,
-                upstream: nextUpstream,
-                info,
-                adapter
-              }
-
-              ctx.followups.set(node, existing)
-            }
-
-            existing.fields.push(field)
-            continue
-          }
-
-          if (s.arguments.length > 0) {
-            value += this.#buildArguments(queryInfo, ctx, s)
-          }
-
-          if (s.selectionSet) {
-            value += ` ${this.#buildSelectionSet(queryInfo, ctx, s, selectionSchemaNode)}`
-          }
-        }
-
-        set.push(value)
-      }
-    }
-
-    ctx.path.pop()
-
-    if (Array.isArray(keyFields)) {
-      for (let i = 0; i < keyFields.length; ++i) {
-        const keyFieldName = keyFields[i]
-
-        if (!set.includes(keyFieldName)) {
-          set.push(keyFieldName)
-        }
-      }
-    }
-
-    return `{ ${set.join(', ')} }`
-  }
+  return mergeAt
 }
 
 function addFieldToType (type, field, subgraph) {
